requirements.txt
# requirements.txt

django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
pymongo==3.12.3
pytz
sqlparse==0.2.4
django-cors-headers==3.13.0
dnspython==2.7.0
gunicorn